# refactor: convert macros in `query_tests` to functions