@crepererum 2021-09-22
Measurements above are taking by:

```console
$ rm -rf target/debug/build/query_tests-* target/debug/deps/query_tests-* target/debug/incremental/query_tests-*
$ cd query_tests/
$ RUSTFLAGS="-Zself-profile" cargo +nightly build
$ cd ..
$ summarize summarize query_tests-<PID>.mm_profdata
```

Also see <https://blog.rust-lang.org/inside-rust/2020/02/25/intro-rustc-self-profile.html>.